Edit attributes dialog: add chattr/chflags stuff

+++ WinPort/sudo.h
@@ -73,8 +73,8 @@ extern "C" {
  __attribute__ ((visibility("default"))) ssize_t sdc_flistxattr(int fd, char *namebuf, size_t size);
  __attribute__ ((visibility("default"))) ssize_t sdc_fgetxattr(int fd, const char *name,void *value, size_t size);
  __attribute__ ((visibility("default"))) int sdc_fsetxattr(int fd, const char *name, const void *value, size_t size, int flags);
- __attribute__ ((visibility("default"))) int sdc_fs_flags_get(const char *path, int *flags);
- __attribute__ ((visibility("default"))) int sdc_fs_flags_set(const char *path, int flags);
+ __attribute__ ((visibility("default"))) int sdc_fs_flags_get(const char *path, unsigned long *flags);
+ __attribute__ ((visibility("default"))) int sdc_fs_flags_set(const char *path, unsigned long flags);
 #ifdef __cplusplus
 }
